Moved 3sat to examples
